requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
cryptography==41.0.7
flask-cors==4.0.0
bcrypt==4.1.2